You typo'ed the word doctype
It'd be nice to see the whole declaration cleaned up and be a valid HTML page such as the one described here: http://stackoverflow.com/a/10963261/633438

However... this code looks generated and there's no instructions on how to develop on this repository. So, I've tried to not over-complicate my PR. Just consider that this is the very first thing in your Getting Started introduction. You'd convert people to JSPM much more readily if one could copy/paste your example and run with it from the documentation.